Add "touch-action" for mobile devices
On a mobile browser, touching and dragging scrolls the screen, but
we use that action for all drawing actions.  We can disable panning
with the CSS "touch-action" property.  We still want to allow
pinch-zoom to work, so that the user can't get trapped if the image
panel fills the entire screen.