Properly set hi-res mode byte
The file type note for $08/0000-3fff specifies values to set in the
first byte of the first screen hole at +120.  We now set it to color
or B&W depending on the value of the "monochrome" checkbox when the
file was saved.  When an image is opened, if our signature is
present, we set the color mode appropriately.